elixir_pre_commit v0.1.0 mix pre_commit

This file contains the functions that will be run when mix pre_commit is run. (we run it in the script in the pre-commit file in your .git/hooks directory but you can run it yourself if you want to see the output without committing).

In here we just run all of the mix commands that you have put in your config file, and if they’re succesful, print a success message to the the terminal, and if they fail we halt the process with a 1 error code ( meaning that the command has failed), which will trigger the commit to stop, and print the error message to the terminal.

Functions

run(_)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.